requirements.txt
anthropic
python-dotenv
streamlit
matplotlib
groq